+++ egui/src/widgets/selected_label.rs
@@ -42,7 +42,7 @@ impl Widget for SelectableLabel {
 
         let visuals = ui.style().interact(&response);
 
-        if selected || response.hovered {
+        if selected || response.hovered() {
             let rect = rect.expand(visuals.expansion);
             let fill = if selected {
                 ui.style().visuals.selection.bg_fill