+++ egui_demo_lib/src/apps/demo/widgets.rs
@@ -146,7 +146,7 @@ impl Widgets {
         ui.horizontal(|ui| {
             ui.label("Single line text input:");
             let response = ui.text_edit_singleline(&mut self.single_line_text_input);
-            if response.lost_kb_focus {
+            if response.lost_kb_focus() {
                 // The user pressed enter.
             }
         });